oneDNN SparseMatrixMatMul removing extraneous test file, fixing copyright years and comments
