+++ tensorflow/python/kernel_tests/linalg/sparse/BUILD
@@ -105,10 +105,15 @@ cuda_py_strict_test(
     ],
 )
 
-py_strict_library(
-    name = "csr_sparse_matrix_dense_mat_mul_grad_test_base",
-    srcs = ["csr_sparse_matrix_dense_mat_mul_grad_test_base.py"],
-    srcs_version = "PY3",
+cuda_py_strict_test(
+    name = "csr_sparse_matrix_dense_mat_mul_grad_test",
+    size = "medium",
+    srcs = ["csr_sparse_matrix_dense_mat_mul_grad_test.py"],
+    main = "csr_sparse_matrix_dense_mat_mul_grad_test.py",
+    shard_count = 50,
+    tags = [
+        "no_cuda_asan",  # TODO(b/190824595)
+    ],
     deps = [
         "//tensorflow/python/framework:ops",
         "//tensorflow/python/framework:test_lib",
@@ -124,25 +129,24 @@ py_strict_library(
 tf_py_test(
     name = "csr_sparse_matrix_dense_mat_mul_onednn_grad_test",
     size = "medium",
-    srcs = ["csr_sparse_matrix_dense_mat_mul_onednn_grad_test.py"],
-    main = "csr_sparse_matrix_dense_mat_mul_onednn_grad_test.py",
-    shard_count = 50,
-    deps = [
-        ":csr_sparse_matrix_dense_mat_mul_grad_test_base",
-    ],
-)
-
-cuda_py_strict_test(
-    name = "csr_sparse_matrix_dense_mat_mul_grad_test",
-    size = "medium",
     srcs = ["csr_sparse_matrix_dense_mat_mul_grad_test.py"],
+    env = {
+        "TF_ENABLE_ONEDNN_SPMM": "1",
+    },
     main = "csr_sparse_matrix_dense_mat_mul_grad_test.py",
     shard_count = 50,
     tags = [
         "no_cuda_asan",  # TODO(b/190824595)
     ],
     deps = [
-        ":csr_sparse_matrix_dense_mat_mul_grad_test_base",
+        "//tensorflow/python/framework:ops",
+        "//tensorflow/python/framework:test_lib",
+        "//tensorflow/python/ops:array_ops",
+        "//tensorflow/python/ops:gradient_checker",
+        "//tensorflow/python/ops/linalg/sparse:sparse_csr_matrix_grad",
+        "//tensorflow/python/ops/linalg/sparse:sparse_csr_matrix_ops",
+        "//tensorflow/python/platform:client_testlib",
+        "//tensorflow/python/platform:tf_logging",
     ],
 )
